añadido mostrar retenciones seccion finanzas distribuidores

+++ database/seeds/agregarComisionesSeeder.php
@@ -11,38 +11,40 @@ class agregarComisionesSeeder extends Seeder
      */
     public function run()
     {
-        
-        if (($gestor = fopen('public/temp/comisiones.csv', "r")) !== FALSE) {
-            fgetcsv($gestor, 1000, ",");
-            while (($vars = fgetcsv($gestor, 1000, ",")) !== FALSE) {
-               $numero = str_replace('"', '',$vars[0]);
-               $periodo = str_replace('"', '',$vars[2]);
-               try{
-                    $simc = \DB::table('simcards')->where('numero', '=',$numero)->first();
-                    if($simc == null){
-                        $ICC = \DB::table('simcards')->select('ICC')->orderBy(\DB::raw('ICC*1'))->first();
-                        $ICC = $ICC->ICC - 1;
-                        $simc = \App\Simcard::create([
-                         'ICC' => $ICC,
-                         'numero' => $numero,
-                         'fecha_vencimiento' => null,
-                         'fecha_activacion' =>  null,
-                         'nombreSubdistribuidor' => 'SIN ASIGNAR',
-                         'tipo' => 1,
-                         'paquete' => 0,
-                         'fecha_entrega' => null
-                         ]);
-                    }
-                    \App\Comision::create([
-                             'ICC' => $simc->ICC,
-                             'valor' => $vars[1],
-                             'periodo' => $periodo,
-                             ]);   
-               }catch(Exception $e){
-               }
+        $this->command->info('Subir archivo comisiones...');
+        if (file_exists('public/temp/comisiones.csv')) {
+            if (($gestor = fopen('public/temp/comisiones.csv', "r")) !== FALSE) {
+                fgetcsv($gestor, 1000, ",");
+                while (($vars = fgetcsv($gestor, 1000, ",")) !== FALSE) {
+                   $numero = str_replace('"', '',$vars[0]);
+                   $periodo = str_replace('"', '',$vars[2]);
+                   try{
+                        $simc = \DB::table('simcards')->where('numero', '=',$numero)->first();
+                        if($simc == null){
+                            $ICC = \DB::table('simcards')->select('ICC')->orderBy(\DB::raw('ICC*1'))->first();
+                            $ICC = $ICC->ICC - 1;
+                            $simc = \App\Simcard::create([
+                             'ICC' => $ICC,
+                             'numero' => $numero,
+                             'fecha_vencimiento' => null,
+                             'fecha_activacion' =>  null,
+                             'nombreSubdistribuidor' => 'SIN ASIGNAR',
+                             'tipo' => 1,
+                             'paquete' => 0,
+                             'fecha_entrega' => null
+                             ]);
+                        }
+                        \App\Comision::create([
+                                 'ICC' => $simc->ICC,
+                                 'valor' => $vars[1],
+                                 'periodo' => $periodo,
+                                 ]);   
+                   }catch(Exception $e){
+                   }
+                }
+                fclose($gestor);
+                unlink('public/temp/comisiones.csv');
             }
-            fclose($gestor);
-            unlink('public/temp/comisiones.csv');
         }
     }
 }
